All model tests work now
