+++ tests/models/test_compound.py
@@ -1,20 +1,9 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, unicode_literals
 
-from scout.models import Compound, Variant
+from .setup_objects import setup_compound, setup_variant
 
 
-def setup_compound(**kwargs):
-  """
-  Setup an Compound object object
-  """
-  compound = Compound(
-    variant = kwargs.get('variant', None),
-    display_name = kwargs.get('display_name', '1_132_A_C'),
-    combined_score = kwargs.get('combined_score', '13'),
-  )
-
-  return compound
   
 def test_compound():
   """
@@ -22,7 +11,7 @@ def test_compound():
   """
   compound = setup_compound()
   
-  assert compound.variant == None
+  assert compound.variant == setup_variant()
   assert compound.display_name == '1_132_A_C'
   assert compound.combined_score == 13
