+++ tests/models/test_individual.py
@@ -1,29 +1,15 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, unicode_literals
 
-from scout.models import Individual
-
-def setup_individual(individual_id='A', display_name='A', sex='1', 
-                    phenotype=1, father='C', mother='B', 
-                    capture_kits=['Nimblegen'], bam_file='path/to/bam'):
-  """
-  Setup an Individual object with the given parameters
-  """
-  individual = Individual(
-                  individual_id = individual_id,
-                  display_name = display_name,
-                  sex = sex,
-                  phenotype = phenotype,
-                  father = father,
-                  mother = mother,
-                  capture_kits = capture_kits,
-                  bam_file = bam_file
-                  )
-  return individual
+from .setup_objects import setup_individual
   
 def test_individual():
-  """docstring for test_individual"""
+  """
+  Test the Individual class
+  """
+  
   individual = setup_individual()
+  
   assert individual.individual_id == 'A'
   assert individual.display_name == 'A'
   assert individual.sex == '1'
